Split main playbook to individual playbooks
There are parts which are specific to our environment. To simplify usage
playbooks have been split to indivual playbooks and the remaining
playbook *playbook.yml* just import theses. Now, people who want to use
this repository just have to comment the import of *provisioning.yml*
playbook (and *monitoring.yml* if the monitoring stack is not wanted).

*docker* and *certificates* have also been included as they are required
for the playbook to work.